% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawExpression.R
\docType{package}
\name{drawExpression}
\alias{drawExpression}
\alias{drawExpression-package}
\alias{drawExpression}
\title{drawExpression}
\usage{
drawExpression(expr, draw.index = FALSE, draw.names = FALSE,
  filename = NULL)
}
\arguments{
\item{expr}{A character vector containing an R expression to be drawn; cannot contain affectation.}

\item{draw.index}{Should index (for R objects such as vector, matrix or list) be drawn on the margin of graphics?}

\item{draw.names}{Should names (if any) be drawn on the margin of objects?}

\item{filename}{The name of a PDF file where the plot will be saved}
}
\value{
nothing
}
\description{
Visualising R syntax through graphics

Draw a graphical representation of the evaluation of an R expression.
}
\examples{
  drawExpression("1:4");
  drawExpression("matrix(1:4, 2)");
  drawExpression("list(1:4, matrix(1:4, 2), 2, 3, 4)");
  drawExpression("c(1, 2, 3:5) > 2");
  
  x <- 1:4
  drawExpression("x");
  drawExpression("sum(x)");
}
\author{
Sylvain Loiseau <sylvain.loiseau@univ-paris13.fr>
}
